% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{calc_svds}
\alias{calc_svds}
\title{Function to Calculate svd of specific mz feature - using output from x,y matrix. Used for feature selection.}
\usage{
calc_svds(matrix)
}
\arguments{
\item{matrix}{xyMatrix as output from \code{create_xyMatrix}}
}
\value{
Vector with svd percentage and squared svd percentage respectively
}
\description{
Function to Calculate svd of specific mz feature - using output from x,y matrix. Used for feature selection.
}
