% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{checkAdductPair}
\alias{checkAdductPair}
\title{Function to check colocalisation of suspected adduct pair}
\usage{
checkAdductPair(MSIobject, mz1, mz2, threshold, method = "M1", topn = 20)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{mz1}{First mz value to run colocalisation against}

\item{mz2}{Second mz value to check colocalisation score above given value}

\item{threshold}{Threshold score to define whether features colocalise}

\item{method}{Method to use for scoring colocalisation (Default = c("M1"))}

\item{topn}{Number of top colocalised features to check (default = 20)}
}
\value{
Boolean whether mz feature colocalise
}
\description{
Function to check colocalisation of suspected adduct pair
}
