% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{create_xyMatrix}
\alias{create_xyMatrix}
\title{Function to create x,y matrix for each feature - for feature selection.}
\usage{
create_xyMatrix(MSIobject, mz)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{mz}{mz channel to generate matrix for}
}
\value{
Data matrix for given mz; rows = y coordinate, cols = x-coordinate in the image
}
\description{
Function to create x,y matrix for each feature - for feature selection.
}
