% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{createDatamatrix}
\alias{createDatamatrix}
\title{Function to create data matrix from MSI object}
\usage{
createDatamatrix(MSIobject, inputNA = T)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{inputNA}{Whether to convert 0's in matrix to NA (default = T)}
}
\value{
dataframe with rows = ? cols = ?
}
\description{
Function to create data matrix from MSI object
}
