% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{error}
\alias{error}
\title{Function to calculate ppm error between mz value and theoretical mass}
\usage{
error(theoretical, measured)
}
\arguments{
\item{theoretical}{theoretical mz of a molecule}

\item{measured}{measured mz}
}
\value{
ppm error
}
\description{
Function to calculate ppm error between mz value and theoretical mass
}
\examples{
error(400, 400.01)

}
