% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMSI_object}
\alias{exampleMSI_object}
\title{Small MSI object pen ink}
\format{
MSI object with 265 features and 25 pixels
}
\usage{
exampleMSI_object
}
\description{
A small MSI object from a sample of pen ink.
}
\keyword{datasets}
