% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{impute_scale}
\alias{impute_scale}
\title{Imputes Missing Values and scales MSI object}
\usage{
impute_scale(MSIobject)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}
}
\value{
MSIobject with missing values imputed and datamatrix scaled.
}
\description{
Imputes Missing Values and scales MSI object
}
