% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{med}
\alias{med}
\title{Function to calculate median of vector excluding 0s and return NA if more 0s than given threshold}
\usage{
med(x, maxNAs = 1)
}
\arguments{
\item{x}{vector to calculate median of}

\item{maxNAs}{Fraction of missing values allowed in vector, if > 0s than \code{maxNAs} an NA is returned}
}
\value{
Median value
}
\description{
Function to calculate median of vector excluding 0s and return NA if more 0s than given threshold
}
\examples{
vec <- c(0,1,0,2,5, 0, 0)
med(vec)

}
