% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{medianNorm}
\alias{medianNorm}
\title{Function for median normalisation}
\usage{
medianNorm(x)
}
\arguments{
\item{x}{vector (relating to pixel spectrum)}
}
\value{
normalised spectrum vector
}
\description{
Function for median normalisation
}
