% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{missing_values_df}
\alias{missing_values_df}
\title{Function to imputes Missing Values into data matrix using knn}
\usage{
missing_values_df(df, k = 10, rowmax = 0.5, colmax = 0.8, maxp = NULL)
}
\arguments{
\item{df}{dataframe with rows = ?, cols = ?}

\item{k}{number of nearest neighbours for knn (default = 10)}

\item{rowmax}{maximum number of missing values in row of df (default = 0.5)}

\item{colmax}{maximum number of missing values in col of df (default = 0.8)}

\item{maxp}{Check \code{impute::impute.knn} (default = NULL)}
}
\value{
df with imputed values - rows = ?, cols = ?
}
\description{
Function to imputes Missing Values into data matrix using knn
}
