% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{myMean}
\alias{myMean}
\title{Function to calculate mean of vector excluding 0s}
\usage{
myMean(x)
}
\arguments{
\item{x}{vector to calculate mean of}
}
\value{
Mean value
}
\description{
Function to calculate mean of vector excluding 0s
}
\examples{
vec <- c(0,1,0,2,5, 0, 0)
myMean(vec)

}
