% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{myMed}
\alias{myMed}
\title{Function to calculate median of vector excluding 0s}
\usage{
myMed(x)
}
\arguments{
\item{x}{vector to calculate median of}
}
\value{
Median value
}
\description{
Function to calculate median of vector excluding 0s
}
\examples{
vec <- c(0,1,0,2,5, 0, 0)
myMed(vec)

}
