% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{normPixelsDF_anatomies}
\alias{normPixelsDF_anatomies}
\title{Function to create dataframe with information about all pixels in several defined anatomical regions of MSI object
Specifically used to compare distribution of feature distributions in different anatomical regions after normalisation}
\usage{
normPixelsDF_anatomies(
  MSIobject,
  pixel_list,
  anatomies,
  normalisation,
  medianDF
)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{pixel_list}{list of pixel vectors from different anatomical regions}

\item{anatomies}{vector of labels for the different anatomical regions pixels belongs to (same order as 'pixel_list')}

\item{normalisation}{Which normalisation step has been applied to 'MSIobject'}

\item{medianDF}{median spectrum to compare to}
}
\value{
Dataframe with headers: pixel, anatomy, mz, intensity (of mz feature), median, FC (to the median spectrum)
}
\description{
Function to create dataframe with information about all pixels in several defined anatomical regions of MSI object
Specifically used to compare distribution of feature distributions in different anatomical regions after normalisation
}
