% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{pixelDF_NormCompare}
\alias{pixelDF_NormCompare}
\title{Function to create dataframe with information about specific pixel in specific anatomical region of MSI object}
\usage{
pixelDF_NormCompare(MSIobject, pixel, anatomy, normalisation, medianDF)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{pixel}{pixel to select spectrum from}

\item{anatomy}{label for the anatomical region pixel belongs to}

\item{normalisation}{Which normalisation step has been applied to 'MSIobject'}

\item{medianDF}{median spectrum to compare to}
}
\value{
Dataframe with headers: pixel, anatomy, mz, intensity (of mz feature), median, FC (to the median spectrum), normalisation
}
\description{
Function to create dataframe with information about specific pixel in specific anatomical region of MSI object
}
