% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardinalExtensionFunctions.R
\name{pixelsDF_NormCompare}
\alias{pixelsDF_NormCompare}
\title{Function to create dataframe with information about all pixels in defined anatomical region of MSI object}
\usage{
pixelsDF_NormCompare(MSIobject, pixels, anatomy, normalisation, medianDF)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{pixels}{pixels to select spectra from}

\item{anatomy}{label for the anatomical region pixels belongs to}

\item{normalisation}{Which normalisation step has been applied to 'MSIobject'}

\item{medianDF}{median spectrum to compare to}
}
\value{
Dataframe with headers: pixel, anatomy, mz, intensity (of mz feature), median, FC (to the median spectrum)
}
\description{
Function to create dataframe with information about all pixels in defined anatomical region of MSI object
}
