% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{ppm_limits}
\alias{ppm_limits}
\title{Function to calculate tolerance m/z values from m/z value and ppm tolerance}
\usage{
ppm_limits(mz, ppm_tol)
}
\arguments{
\item{mz}{mz of interest}

\item{ppm_tol}{ppm tolerance to calculate lower and upper limits}
}
\value{
Vector of lower, upper mz values and mz of interest respectively.
}
\description{
Function to calculate tolerance m/z values from m/z value and ppm tolerance
}
\examples{
ppm_limits(400, 30)

}
