% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{ggplot theme colour}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
ggplot theme colour
}
\examples{
# Use on ggplot object
scale_colour_Publication()

}
\author{
Koundinya Desiraju \url{https://rpubs.com/Koundy/71792}
}
