% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{ggplot theme fill}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{Additional arguments.}
}
\description{
ggplot theme fill
}
\examples{
# Use on ggplot object
scale_fill_Publication()

}
\author{
Koundinya Desiraju \url{https://rpubs.com/Koundy/71792}
}
