% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot theme general}
\usage{
theme_Publication(base_size = 14, base_family = "helvetica")
}
\arguments{
\item{base_size}{Set the text size.}

\item{base_family}{Set the font type.}
}
\description{
ggplot theme general
}
\examples{
# Use on ggplot object
theme_Publication()

}
\author{
Koundinya Desiraju \url{https://rpubs.com/Koundy/71792}
}
